Move out vertex-related rendering logic
